backend/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
numpy==1.24.3
tensorflow==2.13.0
google-api-python-client==2.100.0
shapely==2.0.1
Pillow==10.0.1
scikit-image==0.21.0
aiohttp==3.8.5